pyproject.toml
[project]
name = "fastapi_hash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.0",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "python-jose[criptography,cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "ruff>=0.11.10",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-utils>=0.41.2",
    "taskipy>=1.14.1",
    "uvicorn>=0.34.2",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', 'versions']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
# run = 'fastapi dev main.py'
run = 'uvicorn main:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'
